+++ .gitignore
@@ -1,4 +1,3 @@
-/.phpunit.cache
 /node_modules
 /public/build
 /public/hot
@@ -10,7 +9,7 @@
 .env.backup
 .env.production
 .phpactor.json
-.phpunit.result.cache
+
 Homestead.json
 Homestead.yaml
 npm-debug.log
