+++ README.md
@@ -1,62 +1,145 @@
-<p align="center"><a href="https://laravel.com" target="_blank"><img src="https://raw.githubusercontent.com/laravel/art/master/logo-lockup/5%20SVG/2%20CMYK/1%20Full%20Color/laravel-logolockup-cmyk-red.svg" width="400" alt="Laravel Logo"></a></p>
+# 🏨 Hotel Booking System - Laravel 12
 
-<p align="center">
-<a href="https://github.com/laravel/framework/actions"><img src="https://github.com/laravel/framework/workflows/tests/badge.svg" alt="Build Status"></a>
-<a href="https://packagist.org/packages/laravel/framework"><img src="https://img.shields.io/packagist/dt/laravel/framework" alt="Total Downloads"></a>
-<a href="https://packagist.org/packages/laravel/framework"><img src="https://img.shields.io/packagist/v/laravel/framework" alt="Latest Stable Version"></a>
-<a href="https://packagist.org/packages/laravel/framework"><img src="https://img.shields.io/packagist/l/laravel/framework" alt="License"></a>
-</p>
+Hệ thống đặt phòng khách sạn được xây dựng bằng Laravel 12 với đầy đủ tính năng quản lý booking, phòng, khách hàng và đánh giá.
 
-## About Laravel
+## ✨ Tính năng chính
 
-Laravel is a web application framework with expressive, elegant syntax. We believe development must be an enjoyable and creative experience to be truly fulfilling. Laravel takes the pain out of development by easing common tasks used in many web projects, such as:
+### 🛏️ Quản lý phòng
+- Quản lý loại phòng và phòng
+- Hình ảnh và mô tả chi tiết
+- Trạng thái phòng (trống, đã đặt, bảo trì)
+- Tìm kiếm và lọc phòng
 
-- [Simple, fast routing engine](https://laravel.com/docs/routing).
-- [Powerful dependency injection container](https://laravel.com/docs/container).
-- Multiple back-ends for [session](https://laravel.com/docs/session) and [cache](https://laravel.com/docs/cache) storage.
-- Expressive, intuitive [database ORM](https://laravel.com/docs/eloquent).
-- Database agnostic [schema migrations](https://laravel.com/docs/migrations).
-- [Robust background job processing](https://laravel.com/docs/queues).
-- [Real-time event broadcasting](https://laravel.com/docs/broadcasting).
+### 📅 Hệ thống đặt phòng
+- Đặt phòng trực tuyến
+- Quản lý check-in/check-out
+- Tính toán giá phòng tự động
+- Gia hạn và hủy đặt phòng
+- Hệ thống ghi chú booking
 
-Laravel is accessible, powerful, and provides tools required for large, robust applications.
+### 👥 Quản lý khách hàng
+- Đăng ký và đăng nhập
+- Hồ sơ khách hàng
+- Lịch sử đặt phòng
+- Đánh giá và bình luận
 
-## Learning Laravel
+### ⭐ Hệ thống đánh giá
+- Đánh giá sao và bình luận
+- Đánh giá chi tiết (vệ sinh, tiện nghi, vị trí...)
+- Quản lý đánh giá (admin)
+- Hiển thị đánh giá trung bình
 
-Laravel has the most extensive and thorough [documentation](https://laravel.com/docs) and video tutorial library of all modern web application frameworks, making it a breeze to get started with the framework.
+### 🔧 Admin Panel
+- Dashboard thống kê
+- Quản lý booking
+- Quản lý phòng và loại phòng
+- Quản lý khách hàng
+- Quản lý đánh giá
+- Hệ thống ghi chú
 
-You may also try the [Laravel Bootcamp](https://bootcamp.laravel.com), where you will be guided through building a modern Laravel application from scratch.
+## 🚀 Cài đặt
 
-If you don't feel like reading, [Laracasts](https://laracasts.com) can help. Laracasts contains thousands of video tutorials on a range of topics including Laravel, modern PHP, unit testing, and JavaScript. Boost your skills by digging into our comprehensive video library.
+### Yêu cầu hệ thống
+- PHP 8.2+
+- Laravel 12
+- MySQL/PostgreSQL
+- Composer
+- Node.js & NPM
 
-## Laravel Sponsors
+### Cài đặt
+```bash
+# Clone repository
+git clone [repository-url]
+cd hotel_booking_laravel
 
-We would like to extend our thanks to the following sponsors for funding Laravel development. If you are interested in becoming a sponsor, please visit the [Laravel Partners program](https://partners.laravel.com).
+# Cài đặt dependencies
+composer install
+npm install
 
-### Premium Partners
+# Tạo file .env
+cp .env.example .env
 
-- **[Vehikl](https://vehikl.com/)**
-- **[Tighten Co.](https://tighten.co)**
-- **[Kirschbaum Development Group](https://kirschbaumdevelopment.com)**
-- **[64 Robots](https://64robots.com)**
-- **[Curotec](https://www.curotec.com/services/technologies/laravel/)**
-- **[DevSquad](https://devsquad.com/hire-laravel-developers)**
-- **[Redberry](https://redberry.international/laravel-development/)**
-- **[Active Logic](https://activelogic.com)**
+# Tạo key ứng dụng
+php artisan key:generate
 
-## Contributing
+# Chạy migration
+php artisan migrate
 
-Thank you for considering contributing to the Laravel framework! The contribution guide can be found in the [Laravel documentation](https://laravel.com/docs/contributions).
+# Chạy seeder (nếu có)
+php artisan db:seed
 
-## Code of Conduct
+# Build assets
+npm run build
 
-In order to ensure that the Laravel community is welcoming to all, please review and abide by the [Code of Conduct](https://laravel.com/docs/contributions#code-of-conduct).
+# Chạy server
+php artisan serve
+```
 
-## Security Vulnerabilities
+## 📁 Cấu trúc dự án
 
-If you discover a security vulnerability within Laravel, please send an e-mail to Taylor Otwell via [taylor@laravel.com](mailto:taylor@laravel.com). All security vulnerabilities will be promptly addressed.
+```
+hotel_booking_laravel/
+├── app/
+│   ├── Console/Commands/     # Console commands
+│   ├── Http/
+│   │   ├── Controllers/      # Controllers
+│   │   └── Middleware/       # Middleware
+│   ├── Models/               # Eloquent models
+│   ├── Services/             # Business logic
+│   ├── Repositories/         # Data access layer
+│   └── Interfaces/           # Service interfaces
+├── resources/
+│   ├── views/
+│   │   ├── admin/           # Admin views
+│   │   └── client/          # Client views
+│   └── css/                 # Stylesheets
+├── public/
+│   ├── admin/               # Admin assets
+│   └── client/              # Client assets
+└── database/
+    ├── migrations/          # Database migrations
+    └── seeders/             # Database seeders
+```
 
-## License
+## 🔧 Cấu hình
 
-The Laravel framework is open-sourced software licensed under the [MIT license](https://opensource.org/licenses/MIT).
-# hotel_booking_laravel
+### Database
+Cấu hình database trong file `.env`:
+```env
+DB_CONNECTION=mysql
+DB_HOST=127.0.0.1
+DB_PORT=3306
+DB_DATABASE=hotel_booking
+DB_USERNAME=root
+DB_PASSWORD=
+```
+
+### Mail
+Cấu hình email để gửi thông báo:
+```env
+MAIL_MAILER=smtp
+MAIL_HOST=smtp.mailtrap.io
+MAIL_PORT=2525
+MAIL_USERNAME=your_username
+MAIL_PASSWORD=your_password
+MAIL_ENCRYPTION=tls
+```
+
+## 📚 Tài liệu
+
+Xem thêm tài liệu chi tiết trong thư mục `docs/`:
+- [Hệ thống ghi chú booking](docs/BOOKING_NOTES_SYSTEM.md)
+- [Hướng dẫn kiến trúc](docs/architecture_guidelines.md)
+- [Hướng dẫn interface](docs/interface_contract_guide.md)
+
+## 🤝 Đóng góp
+
+1. Fork dự án
+2. Tạo feature branch (`git checkout -b feature/AmazingFeature`)
+3. Commit thay đổi (`git commit -m 'Add some AmazingFeature'`)
+4. Push to branch (`git push origin feature/AmazingFeature`)
+5. Tạo Pull Request
+
+## 📄 License
+
+Dự án này được phát hành dưới [MIT License](LICENSE).
