--- app/Console/Kernel.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace App\Console;
-
-use Illuminate\Console\Scheduling\Schedule;
-use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
-
-class Kernel extends ConsoleKernel
-{
-    /**
-     * Define the application's command schedule.
-     */
-    protected function schedule(Schedule $schedule): void
-    {
-        // Tạo ghi chú nhắc nhở check-in/check-out hàng ngày lúc 8:00 sáng
-        $schedule->command('booking-notes:create-reminders')
-            ->dailyAt('08:00')
-            ->appendOutputTo(storage_path('logs/booking-notes-reminders.log'));
-    }
-
-    /**
-     * Register the commands for the application.
-     */
-    protected function commands(): void
-    {
-        $this->load(__DIR__.'/Commands');
-
-        require base_path('routes/console.php');
-    }
-}
