+++ app/Observers/BookingNoteObserver.php
@@ -3,59 +3,59 @@
 namespace App\Observers;
 
 use App\Models\BookingNote;
-use App\Services\AdminNotificationService;
+use App\Interfaces\Services\BookingServiceInterface;
 
 class BookingNoteObserver
 {
-    protected $adminNotificationService;
+    protected $bookingService;
 
-    public function __construct(AdminNotificationService $adminNotificationService)
+    public function __construct(BookingServiceInterface $bookingService)
     {
-        $this->adminNotificationService = $adminNotificationService;
+        $this->bookingService = $bookingService;
     }
 
     /**
      * Handle the BookingNote "created" event.
      */
-    public function created(BookingNote $bookingNote): void
+    public function created(BookingNote $bookingNote)
     {
         // Tạo thông báo admin khi có ghi chú mới
-        $this->adminNotificationService->notifyBookingNoteCreated($bookingNote);
+        // $this->bookingService->notifyBookingNoteCreated($bookingNote);
     }
 
     /**
      * Handle the BookingNote "updated" event.
      */
-    public function updated(BookingNote $bookingNote): void
+    public function updated(BookingNote $bookingNote)
     {
         // Tạo thông báo admin khi ghi chú được cập nhật
-        $this->adminNotificationService->notifyBookingNoteUpdated($bookingNote);
+        // $this->bookingService->notifyBookingNoteUpdated($bookingNote);
     }
 
     /**
      * Handle the BookingNote "deleted" event.
      */
-    public function deleted(BookingNote $bookingNote): void
+    public function deleted(BookingNote $bookingNote)
     {
         // Tạo thông báo admin khi ghi chú bị xóa
-        $this->adminNotificationService->notifyBookingNoteDeleted($bookingNote);
+        // $this->bookingService->notifyBookingNoteDeleted($bookingNote);
     }
 
     /**
      * Handle the BookingNote "restored" event.
      */
-    public function restored(BookingNote $bookingNote): void
+    public function restored(BookingNote $bookingNote)
     {
         // Tạo thông báo admin khi ghi chú được khôi phục
-        $this->adminNotificationService->notifyBookingNoteRestored($bookingNote);
+        // $this->bookingService->notifyBookingNoteRestored($bookingNote);
     }
 
     /**
      * Handle the BookingNote "force deleted" event.
      */
-    public function forceDeleted(BookingNote $bookingNote): void
+    public function forceDeleted(BookingNote $bookingNote)
     {
         // Tạo thông báo admin khi ghi chú bị xóa vĩnh viễn
-        $this->adminNotificationService->notifyBookingNoteForceDeleted($bookingNote);
+        // $this->bookingService->notifyBookingNoteForceDeleted($bookingNote);
     }
 }
