--- guidelines_rule
@@ -1,30 +1,0 @@
-## 1. Tổng quan kiến trúc
-
-Hệ thống được xây dựng theo mô hình phân lớp với các thành phần chính sau:
-
-```
-Client Request → Controller → Service → Repository → Model → Database
-```
-
-### Các lớp chính:
-
-- **Models**: Đại diện cho các đối tượng dữ liệu và quan hệ giữa chúng
-- **Repositories**: Xử lý truy vấn và tương tác với cơ sở dữ liệu
-- **Services**: Chứa logic nghiệp vụ 
-- **Controllers**: Xử lý request và response
-- **Interfaces**: Định nghĩa các contract mà các lớp cần tuân thủ
-
-## 2. Quy tắc và nguyên tắc
-
-### SOLID Principles
-- **Single Responsibility**: Mỗi lớp chỉ có một lý do để thay đổi
-- **Open/Closed**: Mở rộng, không sửa đổi
-- **Liskov Substitution**: Các lớp con phải thay thế được lớp cha
-- **Interface Segregation**: Interface nhỏ, chuyên biệt
-- **Dependency Inversion**: Phụ thuộc vào abstraction, không phụ thuộc vào implementation
-
-### Quy tắc đặt tên
-- **Interface**: `{Tên}Interface` (ví dụ: `UserRepositoryInterface`)
-- **Repository**: `{Tên}Repository` (ví dụ: `UserRepository`)
-- **Service**: `{Tên}Service` (ví dụ: `AuthService`)
-- **Controller**: `{Tên}Controller` (ví dụ: `HotelController`)
