+++ public/admin/css/admin-main.css
@@ -85,34 +85,112 @@ h1, h2, h3, h4, h5, h6, .navbar-brand {
 /* Sidebar Toggle Button */
 .sidebar-toggle {
     position: fixed;
-    top: 20px;
-    left: 20px;
-    z-index: 1001;
-    background: var(--accent-color);
+    top: 50%;
+    left: 280px;
+    transform: translateY(-50%);
+    z-index: 1002;
+    background: linear-gradient(135deg, var(--accent-color) 0%, #a67c5b 100%);
     border: none;
     color: white;
-    padding: 0.5rem;
-    border-radius: 50%;
-    width: 40px;
-    height: 40px;
-    display: none;
+    padding: 0;
+    border-radius: 0 12px 12px 0;
+    width: 24px;
+    height: 60px;
+    display: flex;
     align-items: center;
     justify-content: center;
     cursor: pointer;
-    transition: var(--transition);
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2);
+    transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+    box-shadow: 0 4px 20px rgba(193, 155, 118, 0.3);
+    backdrop-filter: blur(10px);
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    overflow: hidden;
 }
 
-/* Show toggle button on mobile */
+.sidebar-toggle::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: linear-gradient(135deg, rgba(255, 255, 255, 0.1) 0%, transparent 100%);
+    opacity: 0;
+    transition: opacity 0.3s ease;
+}
+
+.sidebar-toggle:hover::before {
+    opacity: 1;
+}
+
+/* Toggle button states */
+.sidebar-toggle {
+    left: 280px;
+}
+
+.sidebar.collapsed ~ .sidebar-toggle {
+    left: 0;
+}
+
+/* Mobile adjustments */
 @media (max-width: 991.98px) {
     .sidebar-toggle {
-        display: flex;
+        left: 280px;
+    }
+    
+    .sidebar.collapsed ~ .sidebar-toggle {
+        left: 0;
     }
 }
 
 .sidebar-toggle:hover {
-    background: #a67c5b;
-    transform: scale(1.1);
+    background: linear-gradient(135deg, #a67c5b 0%, #8b6b4a 100%);
+    transform: translateY(-50%) scale(1.05);
+    box-shadow: 0 8px 25px rgba(193, 155, 118, 0.4);
+    width: 28px;
+}
+
+.sidebar-toggle:active {
+    transform: translateY(-50%) scale(0.95);
+    box-shadow: 0 2px 10px rgba(193, 155, 118, 0.3);
+}
+
+/* Toggle Button Icon */
+.sidebar-toggle i {
+    font-size: 0.9rem;
+    transition: transform 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+    position: relative;
+    z-index: 1;
+}
+
+.sidebar-toggle:hover i {
+    transform: rotate(180deg);
+}
+
+/* Icon direction based on sidebar state */
+.sidebar.collapsed ~ .sidebar-toggle i {
+    transform: rotate(180deg);
+}
+
+.sidebar.collapsed ~ .sidebar-toggle:hover i {
+    transform: rotate(0deg);
+}
+
+/* Toggle Button Animation */
+@keyframes togglePulse {
+    0% {
+        box-shadow: 0 4px 20px rgba(193, 155, 118, 0.3);
+    }
+    50% {
+        box-shadow: 0 4px 20px rgba(193, 155, 118, 0.6);
+    }
+    100% {
+        box-shadow: 0 4px 20px rgba(193, 155, 118, 0.3);
+    }
+}
+
+.sidebar-toggle.animate {
+    animation: togglePulse 2s infinite;
 }
 
 /* Sidebar States */
@@ -125,7 +203,7 @@ h1, h2, h3, h4, h5, h6, .navbar-brand {
     top: 0;
     bottom: 0;
     left: 0;
-    z-index: 1000;
+    z-index: 1001;
     width: 280px;
     transition: var(--transition);
     overflow-y: auto;
@@ -147,6 +225,20 @@ h1, h2, h3, h4, h5, h6, .navbar-brand {
     max-width: none;
 }
 
+/* Mobile adjustments for main content */
+@media (max-width: 991.98px) {
+    .main-content {
+        margin-left: 0;
+        width: 100%;
+        padding-top: 70px; /* Space for topbar */
+    }
+    
+    .main-content.expanded {
+        margin-left: 0;
+        width: 100%;
+    }
+}
+
 .main-content.expanded {
     margin-left: 0;
     width: 100%;
@@ -160,7 +252,7 @@ h1, h2, h3, h4, h5, h6, .navbar-brand {
     height: 70px;
     background-color: #fff;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-    z-index: 999;
+    z-index: 1000;
     display: flex;
     align-items: center;
     padding: 0 2rem;
@@ -171,6 +263,21 @@ h1, h2, h3, h4, h5, h6, .navbar-brand {
 .topbar.expanded {
     left: 0;
     width: 100%;
+}
+
+/* Mobile topbar adjustments */
+@media (max-width: 991.98px) {
+    .topbar {
+        left: 0;
+        width: 100%;
+        padding-left: 2rem;
+    }
+    
+    .topbar.expanded {
+        left: 0;
+        width: 100%;
+        padding-left: 2rem;
+    }
 }
 
 /* Page Header */
