+++ resources/views/admin/bookings/index.blade.php
@@ -34,9 +34,11 @@
                             <option value="">Tất cả</option>
                             <option value="pending" {{ $status == 'pending' ? 'selected' : '' }}>Chờ xác nhận</option>
                             <option value="confirmed" {{ $status == 'confirmed' ? 'selected' : '' }}>Đã xác nhận</option>
+                            <option value="checked_in" {{ $status == 'checked_in' ? 'selected' : '' }}>Đã nhận phòng</option>
+                            <option value="checked_out" {{ $status == 'checked_out' ? 'selected' : '' }}>Đã trả phòng</option>
                             <option value="completed" {{ $status == 'completed' ? 'selected' : '' }}>Hoàn thành</option>
                             <option value="cancelled" {{ $status == 'cancelled' ? 'selected' : '' }}>Đã hủy</option>
-                            <option value="no-show" {{ $status == 'no-show' ? 'selected' : '' }}>Không đến</option>
+                            <option value="no_show" {{ $status == 'no_show' ? 'selected' : '' }}>Không đến</option>
                         </select>
                     </div>
                 </form>
@@ -75,15 +77,12 @@
                                 <span class="badge bg-{{ 
                                     $booking->status == 'pending' ? 'warning' : 
                                     ($booking->status == 'confirmed' ? 'primary' : 
+                                    ($booking->status == 'checked_in' ? 'info' :
+                                    ($booking->status == 'checked_out' ? 'secondary' :
                                     ($booking->status == 'completed' ? 'success' : 
-                                    ($booking->status == 'no-show' ? 'dark' : 'danger'))) 
+                                    ($booking->status == 'no_show' ? 'dark' : 'danger'))))) 
                                 }}">
-                                    {{ 
-                                        $booking->status == 'pending' ? 'Chờ xác nhận' : 
-                                        ($booking->status == 'confirmed' ? 'Đã xác nhận' : 
-                                        ($booking->status == 'completed' ? 'Hoàn thành' : 
-                                        ($booking->status == 'no-show' ? 'Không đến' : 'Đã hủy'))) 
-                                    }}
+                                    {{ $booking->status_text }}
                                 </span>
                             </td>
                             <td>
