--- storage/framework/testing/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
