--- tests/TestCase.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Tests;
-
-use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
-
-abstract class TestCase extends BaseTestCase
-{
-    //
-}
